pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipecat-extension"
version = "0.1.0"
description = "A test project using pipecat"
authors = [
    {name = "James Bower", email = "james@example.com"},
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pipecat-ai[local,silero]>=0.0.90",
    "python-dotenv>=1.1.1",
    "websockets>=15.0.1",
    "jsonpatch>=1.32",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.13.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests", "spec", "dev"]
